+++ unreal/MicromegasTelemetrySink/Private/InsertStreamRequest.h
@@ -10,3 +10,4 @@ TArray<uint8> FormatInsertLogStreamRequest(const MicromegasTracing::LogStream& s
 TArray<uint8> FormatInsertMetricStreamRequest(const MicromegasTracing::MetricStream& stream);
 TArray<uint8> FormatInsertThreadStreamRequest(const MicromegasTracing::ThreadStream& stream);
 TArray<uint8> FormatInsertNetStreamRequest(const MicromegasTracing::NetStream& stream);
+TArray<uint8> FormatInsertImageStreamRequest(const MicromegasTracing::ImageStream& stream);