+++ unreal/MicromegasTracing/Public/MicromegasTracing/ImageMetadata.h
@@ -1,0 +1,18 @@
+#pragma once
+//
+//  MicromegasTracing/ImageMetadata.h
+//
+#include "MicromegasTracing/ImageEvents.h"
+#include "MicromegasTracing/QueueMetadata.h"
+
+namespace MicromegasTracing
+{
+	template <>
+	struct GetEventMetadata<ImageEvent>
+	{
+		UserDefinedType operator()()
+		{
+			return UserDefinedType(TEXT("ImageEvent"), 0, false, {});
+		}
+	};
+} // namespace MicromegasTracing